+++ src/rp-download/CurlDownloader.cpp
@@ -2,28 +2,160 @@
  * ROM Properties Page shell extension. (rp-download)                      *
  * CurlDownloader.cpp: libcurl-based file downloader.                      *
  *                                                                         *
- * Copyright (c) 2016-2024 by David Korth.                                 *
+ * Copyright (c) 2016-2025 by David Korth.                                 *
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
 #include "stdafx.h"
 #include "CurlDownloader.hpp"
 
-// C++ STL classes.
-using std::string;
+// C includes
+#include "ctypex.h"
 
-// cURL for network access.
-#include <curl/curl.h>
+// C++ STL classes
+#include <mutex>
+using std::string;
+using std::unique_ptr;
+using std::tstring;
+
+// We're opening libcurl dynamically, so we need to define all
+// cURL function prototypes and definitions here.
+#include "curl-mini.h"
+
+#ifndef _WIN32
+// Unix dlopen()
+#  include <dlfcn.h>
+typedef void *HMODULE;
+// T2U8() is a no-op.
+#  define T2U8(str) (str)
+#else /* _WIN32 */
+// Windows LoadLibrary()
+#  include "libwin32common/RpWin32_sdk.h"
+#  include "libwin32common/MiniU82T.hpp"
+#  include "tcharx.h"
+#  define dlsym(handle, symbol)	((void*)GetProcAddress(handle, symbol))
+#  define dlclose(handle)	FreeLibrary(handle)
+using LibWin32Common::T2U8;
+#endif /* _WIN32 */
+
 
 namespace RpDownload {
 
+class HMODULE_deleter
+{
+public:
+	typedef HMODULE pointer;
+
+	void operator()(HMODULE hModule)
+	{
+		if (hModule) {
+			dlclose(hModule);
+		}
+	}
+};
+static unique_ptr<HMODULE, HMODULE_deleter> libcurl_dll;
+static std::once_flag curl_once_flag;
+
+// Function pointer macro
+#define DEF_STATIC_FUNCPTR(f) static __typeof__(f) * p##f = nullptr
+#define DEF_FUNCPTR(f) __typeof__(f) * p##f = nullptr
+#define LOAD_FUNCPTR(f) p##f = (__typeof__(f)*)dlsym(libcurl_dll.get(), #f)
+
+DEF_STATIC_FUNCPTR(curl_slist_append);
+DEF_STATIC_FUNCPTR(curl_slist_free_all);
+DEF_STATIC_FUNCPTR(curl_getdate);
+DEF_STATIC_FUNCPTR(curl_version_info);
+
+DEF_STATIC_FUNCPTR(curl_easy_init);
+DEF_STATIC_FUNCPTR(curl_easy_setopt);
+DEF_STATIC_FUNCPTR(curl_easy_perform);
+DEF_STATIC_FUNCPTR(curl_easy_cleanup);
+DEF_STATIC_FUNCPTR(curl_easy_getinfo);
+
+/**
+ * Initialize libcurl.
+ * Called by std::call_once().
+ *
+ * Check if libcurl_dll is nullptr afterwards.
+ */
+static void init_curl_once(void)
+{
+	// Open libcurl.
+#ifdef _WIN32
+	libcurl_dll.reset(LoadLibraryEx(_T("libcurl.dll"), nullptr, LOAD_LIBRARY_SEARCH_SYSTEM32 | LOAD_LIBRARY_SEARCH_APPLICATION_DIR));
+#else /* !_WIN32 */
+	// TODO: Consistently use either RTLD_NOW or RTLD_LAZY.
+	// Maybe make it a CMake option?
+	// TODO: Check for ABI version 3 too?
+	// Version 4 was introduced with cURL v7.16.0 (October 2006),
+	// but Debian arbitrarily kept it at version 3.
+	// Reference: https://daniel.haxx.se/blog/2024/10/30/eighteen-years-of-abi-stability/
+	libcurl_dll.reset(dlopen("libcurl.so.4", RTLD_LOCAL | RTLD_NOW));
+#endif /* _WIN32 */
+
+	if (!libcurl_dll) {
+		// Could not open libcurl.
+		return;
+	}
+
+	// curl_global_init() is only needed once.
+	DEF_FUNCPTR(curl_global_init);
+
+	// Load all of the function pointers.
+	LOAD_FUNCPTR(curl_global_init);
+	LOAD_FUNCPTR(curl_slist_append);
+	LOAD_FUNCPTR(curl_slist_free_all);
+	LOAD_FUNCPTR(curl_getdate);
+	LOAD_FUNCPTR(curl_version_info);
+
+	LOAD_FUNCPTR(curl_easy_init);
+	LOAD_FUNCPTR(curl_easy_setopt);
+	LOAD_FUNCPTR(curl_easy_perform);
+	LOAD_FUNCPTR(curl_easy_cleanup);
+	LOAD_FUNCPTR(curl_easy_getinfo);
+
+	if (!pcurl_global_init ||
+	    !pcurl_slist_append ||
+	    !pcurl_slist_free_all ||
+	    !pcurl_getdate ||
+	    !pcurl_easy_init ||
+	    !pcurl_easy_setopt ||
+	    !pcurl_easy_perform ||
+	    !pcurl_easy_cleanup ||
+	    !pcurl_easy_getinfo)
+	{
+		// At least one symbol is missing.
+		libcurl_dll.reset();
+		return;
+	}
+
+	// Initialize cURL.
+	CURLcode ret = pcurl_global_init(CURL_GLOBAL_ALL);
+	if (ret != CURLE_OK) {
+		// Something failed...
+		libcurl_dll.reset();
+	}
+}
+
+CurlDownloader::CurlDownloader()
+{
+	std::call_once(curl_once_flag, init_curl_once);
+	// FIXME: Set a flag if initialization fails.
+}
+
 CurlDownloader::CurlDownloader(const TCHAR *url)
 	: super(url)
-{}
+{
+	std::call_once(curl_once_flag, init_curl_once);
+	// FIXME: Set a flag if initialization fails.
+}
 
 CurlDownloader::CurlDownloader(const tstring &url)
 	: super(url)
-{}
+{
+	std::call_once(curl_once_flag, init_curl_once);
+	// FIXME: Set a flag if initialization fails.
+}
 
 /**
  * Internal cURL data write function.
@@ -139,11 +271,29 @@ size_t CurlDownloader::parse_header(char *ptr, size_t size, size_t nitems, void 
 		mtime_str[val_len] = 0;
 
 		// Parse the modification time.
-		curlDL->m_mtime = curl_getdate(mtime_str, nullptr);
+		curlDL->m_mtime = pcurl_getdate(mtime_str, nullptr);
 	}
 
 	// Continue processing.
 	return len;
+}
+
+/**
+ * Get the name of the IDownloader implementation.
+ * @return Name
+ */
+const TCHAR *CurlDownloader::name(void) const
+{
+	return _T("CurlDownloader");
+}
+
+/**
+ * Is this IDownloader object usable?
+ * @return True if it's usable; false if it's not.
+ */
+bool CurlDownloader::isUsable(void) const
+{
+	return ((bool)libcurl_dll);
 }
 
 /**
@@ -162,7 +312,7 @@ int CurlDownloader::download(void)
 	m_mtime = -1;
 
 	// Initialize cURL.
-	CURL *curl = curl_easy_init();
+	CURL *curl = pcurl_easy_init();
 	if (!curl) {
 		// Could not initialize cURL.
 		return -ENOMEM;	// TODO: Better error?
@@ -174,71 +324,110 @@ int CurlDownloader::download(void)
 	// TODO: Send a HEAD request first?
 
 	// Set options for curl's "easy" mode.
-	curl_easy_setopt(curl, CURLOPT_URL, m_url.c_str());
-	curl_easy_setopt(curl, CURLOPT_NOPROGRESS, true);
+	pcurl_easy_setopt(curl, CURLOPT_URL, T2U8(m_url).c_str());
+	pcurl_easy_setopt(curl, CURLOPT_NOPROGRESS, true);
 	// Fail on HTTP errors. (>= 400)
-	curl_easy_setopt(curl, CURLOPT_FAILONERROR, true);
+	pcurl_easy_setopt(curl, CURLOPT_FAILONERROR, true);
 	// Redirection is required for https://amiibo.life/nfc/%08X-%08X
-	curl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, true);
-	curl_easy_setopt(curl, CURLOPT_MAXREDIRS, 8L);
+	pcurl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, true);
+	pcurl_easy_setopt(curl, CURLOPT_MAXREDIRS, 8L);
 	// Request file modification time.
 	// NOTE: Probably not needed for http...
-	curl_easy_setopt(curl, CURLOPT_FILETIME, 1L);
+	pcurl_easy_setopt(curl, CURLOPT_FILETIME, 1L);
 
 	if (m_if_modified_since >= 0) {
 		// Add an "If-Modified-Since" header.
-#if LIBCURL_VERSION_NUM >= 0x073B00
-		curl_easy_setopt(curl, CURLOPT_TIMEVALUE_LARGE, static_cast<curl_off_t>(m_if_modified_since));
-#else /* LIBCURL_VERSION_NUM < 0x073B00 */
-		curl_easy_setopt(curl, CURLOPT_TIMEVALUE, static_cast<long>(m_if_modified_since));
-#endif /* LIBCURL_VERSION_NUM >= 0x073B00 */
-		curl_easy_setopt(curl, CURLOPT_TIMECONDITION, CURL_TIMECOND_IFMODSINCE);
+
+		// Check the cURL version.
+		const curl_version_info_data *const verinfo = pcurl_version_info(CURLVERSION_FIRST);
+		if (verinfo && verinfo->version_num >= CURL_VERSION_BITS(7, 59, 0)) {
+			static_assert(sizeof(curl_off_t) == 8, "sizeof(curl_off_t) != 8");
+			pcurl_easy_setopt(curl, CURLOPT_TIMEVALUE_LARGE, static_cast<curl_off_t>(m_if_modified_since));
+		} else {
+			pcurl_easy_setopt(curl, CURLOPT_TIMEVALUE, static_cast<long>(m_if_modified_since));
+		}
+		pcurl_easy_setopt(curl, CURLOPT_TIMECONDITION, CURL_TIMECOND_IFMODSINCE);
 	}
 
 	// Custom headers
 	struct curl_slist *req_headers = nullptr;
 	if (!m_reqMimeType.empty()) {
 		// Add the "Accept:" header.
-		tstring accept_header = _T("Accept: ");
-		accept_header += m_reqMimeType;
+		string accept_header = "Accept: ";
+		accept_header += T2U8(m_reqMimeType);
 
-		req_headers = curl_slist_append(req_headers, accept_header.c_str());
+		req_headers = pcurl_slist_append(req_headers, accept_header.c_str());
 		
 	}
 
 	if (req_headers) {
-		curl_easy_setopt(curl, CURLOPT_HTTPHEADER, req_headers);
+		pcurl_easy_setopt(curl, CURLOPT_HTTPHEADER, req_headers);
 	}
 
 	// Header and data functions.
-	curl_easy_setopt(curl, CURLOPT_HEADERFUNCTION, parse_header);
-	curl_easy_setopt(curl, CURLOPT_HEADERDATA, this);
-	curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_data);
-	curl_easy_setopt(curl, CURLOPT_WRITEDATA, this);
+	pcurl_easy_setopt(curl, CURLOPT_HEADERFUNCTION, parse_header);
+	pcurl_easy_setopt(curl, CURLOPT_HEADERDATA, this);
+	pcurl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_data);
+	pcurl_easy_setopt(curl, CURLOPT_WRITEDATA, this);
 
 	// Don't use signals. We're running as a plugin, so using
 	// signals might interfere.
-	curl_easy_setopt(curl, CURLOPT_NOSIGNAL, 1);
+	pcurl_easy_setopt(curl, CURLOPT_NOSIGNAL, 1);
 
 	// Set timeouts to ensure we don't take forever.
 	// TODO: User configuration?
 	// - Connect timeout: 2 seconds.
 	// - Total timeout: 10 seconds.
-	curl_easy_setopt(curl, CURLOPT_CONNECTTIMEOUT, 2);
-	curl_easy_setopt(curl, CURLOPT_TIMEOUT, 10);
+	pcurl_easy_setopt(curl, CURLOPT_CONNECTTIMEOUT, 2);
+	pcurl_easy_setopt(curl, CURLOPT_TIMEOUT, 10);
+
+#ifdef _WIN32
+	// On Windows, Allow passing through some cURL environment variables:
+	// - CURL_CA_BUNDLE: CA certificate bundle (since it might not use the system cert store)
+	// - CURL_INSECURE: If not empty or 0, disable certificate verification entirely.
+	LPCTSTR curl_ca_bundle = _tgetenv(_T("CURL_CA_BUNDLE"));
+	if (curl_ca_bundle) {
+		pcurl_easy_setopt(curl, CURLOPT_CAINFO, T2U8(curl_ca_bundle).c_str());
+	} else {
+		// Set the default CA bundle path.
+		TCHAR path[MAX_PATH + 32];
+		SetLastError(ERROR_SUCCESS);	// required for XP
+		DWORD dwResult = GetModuleFileName(HINST_THISCOMPONENT, path, _countof(path));
+		if (dwResult == 0 || dwResult >= _countof(path) || GetLastError() != ERROR_SUCCESS) {
+			// Cannot get the DLL filename.
+			// TODO: Windows XP doesn't SetLastError() if the
+			// filename is too big for the buffer.
+		} else {
+			// Got the DLL filename.
+			// NULL out the filename portion, then append "curl-ca-bundle.crt".
+			TCHAR *const last_slash = _tcsrchr(path, _T('\\'));
+			if (last_slash && last_slash < &path[MAX_PATH]) {
+				_tcscpy(&last_slash[1], _T("curl-ca-bundle.crt"));
+				pcurl_easy_setopt(curl, CURLOPT_CAINFO, T2U8(path).c_str());
+			}
+		}
+	}
+
+	LPCTSTR curl_insecure = _tgetenv(_T("CURL_INSECURE"));
+	if (curl_insecure) {
+		if (curl_insecure[0] != _T('\0') && isdigit_ascii(curl_insecure[0]) && _tcscmp(curl_insecure, _T("0")) != 0) {
+			pcurl_easy_setopt(curl, CURLOPT_SSL_VERIFYPEER, 0);
+		}
+	}
+#endif /* _WIN32 */
 
 	// Set the User-Agent.
-	curl_easy_setopt(curl, CURLOPT_USERAGENT, m_userAgent.c_str());
+	pcurl_easy_setopt(curl, CURLOPT_USERAGENT, T2U8(m_userAgent).c_str());
 
 	// Download the file.
 	int ret;
-	CURLcode res = curl_easy_perform(curl);
+	CURLcode res = pcurl_easy_perform(curl);
 	switch (res) {
 		case CURLE_OK:
 			// If the file is empty, check for a 304.
 			if (m_data.empty() && m_if_modified_since >= 0) {
 				long unmet = 0;
-				if (!curl_easy_getinfo(curl, CURLINFO_CONDITION_UNMET, &unmet) && unmet) {
+				if (!pcurl_easy_getinfo(curl, CURLINFO_CONDITION_UNMET, &unmet) && unmet) {
 					// HTTP 304 Not Modified
 					ret = 304;
 					break;
@@ -259,7 +448,7 @@ int CurlDownloader::download(void)
 			// Check if we have an HTTP response code.
 			// NOTE: GameTDB sometimes returns nothing instead of 404...
 			long response_code = 0;
-			curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &response_code);
+			pcurl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &response_code);
 			if (response_code <= 0) {
 				// No HTTP response code.
 				// TODO: Return a cURL error code and/or message...
@@ -269,8 +458,8 @@ int CurlDownloader::download(void)
 			break;
 	}
 
-	curl_slist_free_all(req_headers);
-	curl_easy_cleanup(curl);
+	pcurl_slist_free_all(req_headers);
+	pcurl_easy_cleanup(curl);
 	if (ret != 0) {
 		return ret;
 	}