+++ src/rp-download/CurlDownloader.hpp
@@ -2,7 +2,7 @@
  * ROM Properties Page shell extension. (rp-download)                      *
  * CurlDownloader.hpp: libcurl-based file downloader.                      *
  *                                                                         *
- * Copyright (c) 2016-2024 by David Korth.                                 *
+ * Copyright (c) 2016-2025 by David Korth.                                 *
  * SPDX-License-Identifier: GPL-2.0-or-later                               *
  ***************************************************************************/
 
@@ -15,7 +15,7 @@ namespace RpDownload {
 class CurlDownloader final : public IDownloader
 {
 public:
-	CurlDownloader() = default;
+	CurlDownloader();
 	explicit CurlDownloader(const TCHAR *url);
 	explicit CurlDownloader(const std::tstring &url);
 
@@ -45,6 +45,18 @@ protected:
 	static size_t parse_header(char *ptr, size_t size, size_t nitems, void *userdata);
 
 public:
+	/**
+	 * Get the name of the IDownloader implementation.
+	 * @return Name
+	 */
+	const TCHAR *name(void) const final;
+
+	/**
+	 * Is this IDownloader object usable?
+	 * @return True if it's usable; false if it's not.
+	 */
+	bool isUsable(void) const final;
+
 	/**
 	 * Download the file.
 	 * @return 0 on success; negative POSIX error code, positive HTTP status code on error.