+++ NEWS.md
@@ -58,6 +58,8 @@
       * Reported by @ldroberts02.
   * TGA: Fix handling of "undefined" alpha channels.
   * TGA: Show the image ID field.
+  * Linux: The i386 version of rp-download was broken because the seccomp
+    syscall whitelist was missing `getuid32()`.
 
 * Other changes:
   * rpcli: Added more colorization for warning messages.
@@ -69,6 +71,10 @@
   * Windows: Files marked as "offline" (e.g. not synchronized with a cloud backup
     provider) will now be considered to be on a "slow" file system, and won't be
     thumbnailed if the option to thumbnail network file systems is unchecked.
+  * Windows: On Windows XP and Server 2003, if libcurl.dll is present in the same
+    directory as rp-download.exe, it will be used instead of WinInet. This allows
+    rom-properties to access servers with more up-to-date security protocols,
+    which isn't normally supported by these older versions of Windows.
 
 ## v2.5.1 (released 2025/05/10)